Fix ParserTest: Null storage error
